WIP file logger
