+++ tests/monitor.ini
@@ -47,6 +47,21 @@ type=logfile
 filename=monitor4.log
 dateformat=iso8601
 
+[file5]
+type=logfileng
+filename=monitor6.log
+rotation_type=time
+when=m
+interval=1
+backup_count=1
+dateformat=iso8601
+
+[file6]
+type=logfileng
+filename=monitor7.log
+rotation_type=size
+max_bytes=1K
+
 [sms]
 type=bulksms
 username=a
